feat: added application of calibration

+++ src/syncro/resolutions.rs
@@ -1,16 +1,16 @@
 pub struct Resoulutions {
-    gain: f32,
-    offset: f32,
+    gain: f64,
+    offset: f64,
 }
 
 impl Resoulutions {
-    pub fn new(gain: f32, offset: f32) -> Self {
+    pub fn new(gain: f64, offset: f64) -> Self {
         Self { gain, offset }
     }
-    pub fn scale_gain(&self, gain: f32) -> u16 {
+    pub fn scale_gain(&self, gain: f64) -> u16 {
         (gain / self.gain).round() as u16
     }
-    pub fn scale_offset(&self, offset: f32) -> u16 {
+    pub fn scale_offset(&self, offset: f64) -> u16 {
         (offset / self.offset).round() as u16
     }
 }
